Analog stick for lift, GreenBot extends mecanum drive
Switched slide controls from d-pad to analog stick
Separated mecanum drive and scoring functions, scoring functions (specific to green bot) are now in GreenBot class
