+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/MarvelsMecanumDrive.java
@@ -16,14 +16,8 @@ public class MarvelsMecanumDrive {
     public MotorEx motor_frontRight;
     public MotorEx motor_backLeft;
     public MotorEx motor_backRight;
-        public MotorEx lift_motor;
-        public MotorEx lift_motor2;
-    public SimpleServo clawLeft;
-    public SimpleServo clawRight;
 
     public MecanumDrive mecanumDrivetrain;
-
-    public boolean clawOpen = true;
 
     private static final double TICKS_PER_REV = DriveConstants.TICKS_PER_REV;
 
@@ -33,28 +27,7 @@ public class MarvelsMecanumDrive {
 
     }
 
-    public void runSlide (double power){
-        lift_motor.set(power);
-        lift_motor2.set(power);
-    }
 
-
-    public void runClaw(){
-        if (clawOpen){
-            clawLeft.setPosition(TeleOpConfig.CLAW_LEFT_OPEN);
-            clawRight.setPosition(TeleOpConfig.CLAW_RIGHT_OPEN);
-        }
-        else{
-            clawLeft.setPosition(TeleOpConfig.CLAW_LEFT_CLOSED);
-            clawRight.setPosition(TeleOpConfig.CLAW_RIGHT_CLOSED);
-        }
-    }
-
-
-
-    public void toggleClaw (){
-        clawOpen = !clawOpen;
-    }
     public void init(HardwareMap hw) {
         //cache the HardwareMap
         this.hw = hw;
@@ -70,18 +43,10 @@ public class MarvelsMecanumDrive {
         motor_frontRight.setInverted(true);
         motor_backRight.setInverted(true);
 
-        clawLeft = new SimpleServo(hw,"claw left",0,180);
-        clawRight = new SimpleServo(hw,"claw right",0,180);
-
-        lift_motor = new MotorEx(hw, "lift motor", TICKS_PER_REV, MAX_RPM);
-        lift_motor2 = new MotorEx(hw, "lift motor 2", TICKS_PER_REV, MAX_RPM);
-
         motor_frontLeft.setZeroPowerBehavior(Motor.ZeroPowerBehavior.BRAKE);
         motor_frontRight.setZeroPowerBehavior(Motor.ZeroPowerBehavior.BRAKE);
         motor_backLeft.setZeroPowerBehavior(Motor.ZeroPowerBehavior.BRAKE);
         motor_backRight.setZeroPowerBehavior(Motor.ZeroPowerBehavior.BRAKE);
-        lift_motor.setZeroPowerBehavior(Motor.ZeroPowerBehavior.BRAKE);
-        lift_motor2.setZeroPowerBehavior(Motor.ZeroPowerBehavior.BRAKE);
