+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOpConfig.java
@@ -7,6 +7,6 @@ public class TeleOpConfig {
     public static double CLAW_LEFT_OPEN = 0.73;
     public static double CLAW_RIGHT_CLOSED = 0.32;
     public static double CLAW_RIGHT_OPEN = 0.15;
-    public static double SLIDE_SPEED = 0.9;
-    public static double SPIN_SPEED = -0.4;
+    // public static double SLIDE_SPEED = 0.9;
+    // public static double SPIN_SPEED = -0.4;
 }
